Update go button `viewBox` to ensure we can never clip
The current go button svg `viewBox` can clip as it results in an
image a few % larger than the containing box. This is fine as long
as the containing box has some extra padding, but in the next
commit we're gonna change that so we need a slightly smaller image.